0.58.4-1 (patches unapplied)
Imported using git-ubuntu import.

+++ lib/Sbuild/Build.pm
@@ -22,12 +22,18 @@
 
 package Sbuild::Build;
 
+use strict;
+use warnings;
+
+use POSIX;
 use Errno qw(:POSIX);
 use Fcntl;
 use File::Basename qw(basename dirname);
 use File::Temp qw(tempdir);
+use FileHandle;
 use GDBM_File;
-use Sbuild qw($devnull binNMU_version version_compare split_version copy isin send_mail debug);
+
+use Sbuild qw($devnull binNMU_version version_compare split_version copy isin send_mail debug df);
 use Sbuild::Base;
 use Sbuild::ChrootSetup qw(update upgrade);
 use Sbuild::ChrootInfoSchroot;
@@ -36,12 +42,6 @@ use Sbuild::Sysconfig qw($version $release_date);
 use Sbuild::Conf;
 use Sbuild::LogBase qw($saved_stdout);
 use Sbuild::Sysconfig;
-
-use strict;
-use warnings;
-use POSIX;
-use FileHandle;
-use File::Temp ();
 
 BEGIN {
     use Exporter ();
@@ -300,7 +300,13 @@ sub run {
 	$session->get('Session Purged') == 1) {
 	$self->log("Not removing build depends: cloned chroot in use\n");
     } else {
-	$self->uninstall_deps();
+	if ($self->get_conf('PURGE_BUILD_DEPS') eq 'always' ||
+	    ($self->get_conf('PURGE_BUILD_DEPS') eq 'successful' &&
+	     $self->get_status() eq 'successful')) {
+	    $self->uninstall_deps();
+	} else {
+	    $self->log("Not removing build depends: as requested\n");
+	}
     }
     $self->remove_srcdep_lock_file();
   cleanup_close:
@@ -644,7 +650,7 @@ sub build {
     $current_usage =~ /^(\d+)/;
     $current_usage = $1;
     if ($current_usage) {
-	my $free = $self->df($dscdir);
+	my $free = df($dscdir);
 	if ($free < 2*$current_usage) {
 	    $self->log("Disc space is propably not enough for building.\n".
 		       "(Source needs $current_usage KB, free are $free KB.)\n");
@@ -784,6 +790,7 @@ sub build {
 	COMMAND => $buildcmd,
 	ENV => $buildenv,
 	USER => $self->get_conf('USERNAME'),
+	SETSID => 1,
 	CHROOT => 1,
 	PRIORITY => 0,
 	DIR => $bdir
@@ -807,8 +814,9 @@ sub build {
 	my $pid = $command->{'PID'};
 	my $signal = ($timed_out > 0) ? "KILL" : "TERM";
 	$self->get('Session')->run_command(
-	    { COMMAND => ['perl', '-e',
-			  "\"kill( \\\"$signal\\\", $pid )\""],
+	    { COMMAND => ['perl',
+			  '-e',
+			  "kill( \"$signal\", -$pid )"],
 	      USER => 'root',
 	      CHROOT => 1,
 	      PRIORITY => 0,
@@ -2301,16 +2309,6 @@ sub unset_removed {
     debug("Removed from removed list: @_\n");
 }
 
-sub df {
-    my $self = shift;
-    my $dir = shift;
-
-    my $df = $Sbuild::Sysconfig::programs{'DF'};
-    my $free = `"$df" $dir | tail -n 1`;
-    my @free = split( /\s+/, $free );
-    return $free[3];
-}
-
 sub fixup_pkgv {
     my $self = shift;
     my $pkgv = shift;
@@ -2461,7 +2459,7 @@ sub chroot_arch {
 
     my $pipe = $self->get('Session')->pipe_command(
 	{ COMMAND => [$self->get_conf('DPKG'),
-		      '--print-installation-architecture'],
+		      '--print-architecture'],
 	  USER => $self->get_conf('USERNAME'),
 	  CHROOT => 1,
 	  PRIORITY => 0,
